move 'pipeK' tests to separate file
